update layout story templates
