+++ src/templates/layouts/storyListing/mockFormattedList.js
@@ -1,0 +1,193 @@
+export const formattedStories = [
+  {
+    id: '1024597e-b397-4b92-a183-9402acfd5d0d',
+    type: 'node--news_story',
+    published: true,
+    title: 'Celebrating our Family Caregivers',
+    image: {
+      id: 'e4db22b0-5802-405f-a526-a3a436bd4219',
+      type: 'media--image',
+      link: {},
+      alt: 'Caregiver Support event',
+      width: 6240,
+      height: 4160,
+      title: '',
+      url: 'https://prod.cms.va.gov/sites/default/files/styles/2_1_large/public/2023-11/IMG_2876.JPG?h=38f1b9a2',
+    },
+    link: '/butler-health-care/stories/celebrating-our-family-caregivers',
+    introText:
+      'Each November, we observe National Family Caregivers Month (NFCM) to celebrate family caregivers nationwide.',
+  },
+  {
+    id: 'cecaf5d7-34d8-42e1-9bc0-c2cc46b38a2a',
+    type: 'node--news_story',
+    published: true,
+    title: "It's Flu Shot Time!",
+    image: {
+      id: 'cd5abddc-ecaa-4a82-bfcb-0eea7477dcc6',
+      type: 'media--image',
+      link: {},
+      alt: "Don't Pass Flu On",
+      width: 5104,
+      height: 2871,
+      title: '',
+      url: 'https://prod.cms.va.gov/sites/default/files/styles/2_1_large/public/2023-09/veteran-flu-prevention-toolkit-bulletinboard-v2.jpg?h=c9522600',
+    },
+    link: '/butler-health-care/stories/its-flu-shot-time',
+    introText:
+      'The Butler VA Health Care System will be offering flu vaccination clinics to Veterans at its New Castle Road campus and community locations this fall. ',
+  },
+  {
+    id: 'ff6969f7-7907-4eae-9571-634153c26bc9',
+    type: 'node--news_story',
+    published: true,
+    title:
+      'This Suicide Prevention Month, the Butler VA Offers Veterans Support',
+    image: {
+      id: '33a81ceb-b5f0-4544-8967-d97214b8edaa',
+      type: 'media--image',
+      link: {},
+      alt: 'Suicide Prevention Month',
+      width: 960,
+      height: 540,
+      title: '',
+      url: 'https://prod.cms.va.gov/sites/default/files/styles/2_1_large/public/2023-09/2-okay-to-ask-for-help-desktop-background.jpg?h=88bff24e',
+    },
+    link: '/butler-health-care/stories/this-suicide-prevention-month-the-butler-va-offers-veterans-support',
+    introText:
+      'Veterans are often the first to help others, but it can sometimes be hard for Veterans to accept or ask for help themselves. This September, throughout Suicide Prevention Month, the Butler VA Health Care System is offering events and resources to help Veterans who are struggling.',
+  },
+  {
+    id: 'd023d3de-a14c-4fb2-b069-1bf411ce3be1',
+    type: 'node--news_story',
+    published: true,
+    title: 'In-car and Body-worn Cameras',
+    image: {
+      id: 'b38b2e4e-5190-4492-9852-b7f5b21d40d5',
+      type: 'media--image',
+      link: {},
+      alt: 'va police officers standing - one male and one female',
+      width: 5934,
+      height: 4470,
+      title: '',
+      url: 'https://prod.cms.va.gov/sites/default/files/styles/2_1_large/public/2023-08/va%20police%20officers%20standing%20-%20one%20male%20and%20one%20female.jpg?h=ec283f65',
+    },
+    link: '/butler-health-care/stories/in-car-and-body-worn-cameras',
+    introText:
+      'The mission of VA’s police force is to protect anyone who comes to VA. ',
+  },
+  {
+    id: 'eb8c6201-1015-4474-a705-a30efb3b9082',
+    type: 'node--news_story',
+    published: true,
+    title: 'High-Tech & High-Touch Health Care',
+    image: {
+      id: '06013955-467f-46fc-a3a6-f5d85abb1fcb',
+      type: 'media--image',
+      link: {},
+      alt: 'telehealth conference',
+      width: 6240,
+      height: 4160,
+      title: '',
+      url: 'https://prod.cms.va.gov/sites/default/files/styles/2_1_large/public/2023-06/IMG_0285.JPG?h=3e68e630',
+    },
+    link: '/butler-health-care/stories/high-tech-high-touch-health-care',
+    introText:
+      'May marked 20 years since the U.S. Department of Veterans Affairs launched its Telehealth Services program office. ',
+  },
+  {
+    id: '5bae8ede-148f-4ef2-8d86-10753b2b9a59',
+    type: 'node--news_story',
+    published: true,
+    title: 'All About Acupuncture  & Chiropractic Care',
+    image: {
+      id: 'a4a15542-d71d-479c-92da-7a5a460ab71d',
+      type: 'media--image',
+      link: {},
+      alt: "Butler VA's Chiropractors and Acupuncturists",
+      width: 4131,
+      height: 3224,
+      title: '',
+      url: 'https://prod.cms.va.gov/sites/default/files/styles/2_1_large/public/2023-05/Feature%20Story_Chiro%20and%20Accu%20Team_2.jpg?h=6c1ae47b',
+    },
+    link: '/butler-health-care/stories/all-about-acupuncture-chiropractic-care',
+    introText:
+      'The Butler VA Health Care System has been providing chiropractic care for over 10 years, and acupuncture services for almost 8. The two health services incorporate Whole Health principles with a common goal of improving Veterans’ overall health, but there are some major differences. ',
+  },
+  {
+    id: 'c2bf92f6-d49a-45e1-9f36-86ffb4176d1c',
+    type: 'node--news_story',
+    published: true,
+    title: '“It’s a Help System”',
+    image: {
+      id: '7c58a640-8580-45c9-96fa-9e1d83c79c4b',
+      type: 'media--image',
+      link: {},
+      alt: 'Army Veteran Calvin Kasparek',
+      width: 3024,
+      height: 2076,
+      title: '',
+      url: 'https://prod.cms.va.gov/sites/default/files/styles/2_1_large/public/2023-04/Health%20Tech_Annie%20App%20Vet_1.jpeg',
+    },
+    link: '/butler-health-care/stories/its-a-help-system',
+    introText:
+      'Annie is a VA messaging service that empowers Veterans to take an active role in their health care. ',
+  },
+  {
+    id: 'fcfcd125-6c2e-4e58-9881-5087d99b51ee',
+    type: 'node--news_story',
+    published: true,
+    title: 'Tele-Audiology Now Available',
+    image: {
+      id: 'ad24c19d-b093-40cb-9e94-55d9aae3ef7e',
+      type: 'media--image',
+      link: {},
+      alt: 'Telehealth appointment',
+      width: 4495,
+      height: 4160,
+      title: '',
+      url: 'https://prod.cms.va.gov/sites/default/files/styles/2_1_large/public/2023-03/Feature%20Story_TeleAudiology%20%281%29_0.jpg?h=e40710da',
+    },
+    link: '/butler-health-care/stories/tele-audiology-now-available',
+    introText:
+      'A new telehealth service is available for Veterans at the Lawrence County Community-based Outpatient Clinic: Tele-Audiology.',
+  },
+  {
+    id: '3460f445-9dc5-4a7e-aeb8-56ac16231e35',
+    type: 'node--news_story',
+    published: true,
+    title: 'Living My Mission, Aspiration, and Purpose',
+    image: {
+      id: 'a5a7a4c7-413e-48b6-a980-d8278d55808d',
+      type: 'media--image',
+      link: {},
+      alt: 'U.S. Army and Army National Guard Veteran Sue Waldoch ',
+      width: 3708,
+      height: 4726,
+      title: '',
+      url: 'https://prod.cms.va.gov/sites/default/files/styles/2_1_large/public/2023-02/Vet%20Chat_Sue.jpg?h=0c60b38b',
+    },
+    link: '/butler-health-care/stories/living-my-mission-aspiration-and-purpose',
+    introText:
+      'VA’s Whole Health approach puts YOU at the center of your care. ',
+  },
+  {
+    id: '6b436546-c3c4-4ec8-8eaa-f11bc520fc91',
+    type: 'node--news_story',
+    published: true,
+    title: 'Equine Therapy Community Partnerships',
+    image: {
+      id: 'c46ec4c9-c47a-47f7-9b49-489a5520feff',
+      type: 'media--image',
+      link: {},
+      alt: 'grandfather and granddaughter riding horses together',
+      width: 2687,
+      height: 1839,
+      title: '',
+      url: 'https://prod.cms.va.gov/sites/default/files/styles/2_1_large/public/2023-01/Equine%20Web%20Story.jpg?h=5eb487b6',
+    },
+    link: '/butler-health-care/stories/equine-therapy-community-partnerships',
+    introText:
+      'Did you know the Butler VA Health Care System currently has 4 Veteran Community Partnerships for equine therapy? ',
+  },
+]
